+++ common/src/main/java/dev/rinchan/keepbutpenalty/DeathStreakAccess.java
@@ -1,0 +1,7 @@
+package dev.rinchan.keepbutpenalty;
+
+public interface DeathStreakAccess {
+    DeathStreakPolicy.State keepButPenalty$getDeathStreak();
+
+    void keepButPenalty$setDeathStreak(DeathStreakPolicy.State state);
+}